// src/models/delegateModel.ts
export interface DelegateForm {
    delegatePass?: string;
    title: string;
    fullName: string;
    jobTitle: string;
    companyName: string;
    email: string;
    mobile: string;
    industry: string;
    country: string;
    role: string;
    solutionLookingFor: string;
    numberOfEmployees: string;
    funding: string;
    implementationTime: string;
    referFullName?: string;
    referEmail?: string;
    referCompanyName?: string;
    referPhoneNumber?: string;
    referJobTitle?: string;
    confirmation:[];
  }
  